Reorganize tests Put all tests in __tests__ directory
